feat: fixed not working callbackGetThreadContext and added callbackCyclesPassed
callbackGetThreadContext  was checking uninitialized ctx structure thus it wasnt doing anything valid.

added callbackCyclesPassed which checks how many times the cycles delta in the function was higher than number x (60000). If it was higher more than or equal 3 times then the function assumes that there was a debugger. (It doesn't check anything after the 3 times barrier was passed)